backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.3.2
# Authentication & Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.3.0

